@@ -40,6 +40,7 @@
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/plugin-proposal-decorators": "^7.8.3",
     "@babel/plugin-transform-async-to-generator": "^7.8.3",
+    "@babel/plugin-transform-react-jsx": "^7.8.3",
     "@babel/polyfill": "^7.8.3",
     "@babel/preset-env": "^7.8.4",
     "@babel/preset-react": "^7.8.3",
@@ -49,16 +50,16 @@
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "copyfiles": "^2.2.0",
-    "coveralls": "^2.13.1",
+    "coveralls": "^3.0.9",
     "cross-env": "^7.0.0",
     "eslint": "^6.8.0",
     "eslint-config-kotori": "^0.3.5",
     "eslint-loader": "^3.0.3",
+    "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.18.3",
     "file-loader": "^5.1.0",
     "friendly-errors-webpack-plugin": "^1.7.0",
     "gh-pages": "^2.2.0",
-    "glob": "^7.1.6",
     "html-webpack-plugin": "^3.2.0",
     "imports-loader": "^0.8.0",
     "istanbul-instrumenter-loader": "^3.0.1",
@@ -70,13 +71,12 @@
     "karma-mocha-reporter": "^2.0.0",
     "karma-sourcemap-loader": "^0.3.5",
     "karma-webpack": "^4.0.2",
-    "minimist": "^1.2.0",
-    "mocha": "^3.5.0",
-    "nock": "^9.0.14",
+    "mocha": "^7.0.1",
     "nodemon": "^2.0.2",
     "null-loader": "^3.0.0",
     "open-browser-webpack-plugin": "^0.0.5",
     "optimize-js-plugin": "^0.0.4",
+    "prettier": "^1.19.1",
     "puppeteer": "^2.1.1",
     "react-hot-loader": "4.12.19",
     "react-test-renderer": "^16.12.0",
@@ -92,8 +92,9 @@
     "webpack-merge": "^4.2.2"
   },
   "dependencies": {
-    "@material-ui/core": "^1.3.0",
-    "@material-ui/icons": "^1.1.0",
+    "@material-ui/core": "^4.9.3",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/styles": "^4.9.0",
     "animation-frame": "^0.3.0",
     "autobind-decorator": "^2.4.0",
     "body-parser": "^1.19.0",
@@ -104,26 +105,25 @@
     "express": "^4.17.1",
     "helmet": "^3.21.2",
     "honoka": "^0.4.17",
-    "inobounce": "^0.1.6",
-    "jss-vendor-prefixer": "^8.0.0",
+    "inobounce": "^0.2.0",
     "jszip": "^3.1.4",
     "moment": "^2.24.0",
     "namespaced-types": "^0.1.2",
     "node-cache": "^5.1.0",
     "pixiv-app-api": "^1.0.4",
     "prop-types": "^15.6.2",
-    "react": "^16.4.1",
+    "react": "^16.8.0",
     "react-document-title": "^2.0.3",
-    "react-dom": "^16.4.1",
+    "react-dom": "^16.8.0",
     "react-event-listener": "^0.6.1",
     "react-ga": "^2.7.0",
-    "react-image": "^1.0.1",
+    "react-image": "^2.4.0",
     "react-intl": "^2.4.0",
     "react-masonry-component": "^6.2.1",
-    "react-md-spinner": "^0.2.5",
+    "react-md-spinner": "^1.0.0",
     "react-modal": "^3.4.5",
-    "react-redux": "^5.0.6",
-    "react-router-dom": "^4.3.1",
+    "react-redux": "^7.2.0",
+    "react-router-dom": "^5.0.0",
     "redux": "^4.0.0",
     "redux-thunk": "^2.3.0",
     "response-time": "^2.3.2",
