@@ -159,9 +159,7 @@ export default class EmojiParser {
       let replaceStr = '';
       for (const series of EmojiParser.emojiSeries) {
         if (match === `(${series.name})`) {
-          replaceStr += `<img src="https://source.pixiv.net/common/images/emoji/${
-            series.id
-          }.png" />`;
+          replaceStr += `<img src="https://source.pixiv.net/common/images/emoji/${series.id}.png" />`;
         }
       }
       return replaceStr === '' ? match : replaceStr;
